Public

Public Functions for Marconi

Index

Marconi.DataNetwork

Marconi.EquationNetwork

Marconi.inputZ

Marconi.inputZ

Marconi.interpolate

Marconi.plotLStabCircle!

Marconi.plotSStabCircle!

Marconi.plotSmithCircle!

Marconi.plotSmithData

Marconi.plotSmithData!

Marconi.plotVSWR!

Marconi.processTouchstoneLine

Marconi.readTouchstone

Marconi.s2t

Marconi.s2z

Marconi.t2s

Marconi.testDelta

Marconi.testK

Marconi.testMagDelta

Marconi.y2s

Marconi.z2s

Marconi.Γ

Marconi.∠

Public Interface

Marconi.DataNetwork — Type.

The base Network type for representing n-port linear networks with characteristic impedance Z0. By default, the network is stored as S-Parameters with the corresponding frequency list.

source

Marconi.EquationNetwork — Type.

The base Network type for representing n-port linear networks with characteristic impedance Z0. The S-Parameters for an EquationNetwork are defined by a function that returns a ports-square matrix and accepts kwargs Z0 and freq.

source

Marconi.inputZ — Method.

inputZ(Zr,Θ,Z0)

Calculates the input impedace of a lossless transmission line of length θ in degrees terminated with Zr. Z0 is optional and defaults to 50.

source

Marconi.inputZ — Method.

inputZ(Γ,Z0)

Calculates the input impedace from complex reflection coefficient Γ. Z0 is optional and defaults to 50.

source

Marconi.interpolate — Method.

interpolate(network,frequencies)

Returns a new network object that contains data from network reinterpolated to fit frequencies.

source

Marconi.plotLStabCircle! — Method.

plotLStabCircle!(sc,network,freq)

Plots the the load stability circle on Smith Chart sc from network at frequency freq.

source

Marconi.plotSStabCircle! — Method.

plotSStabCircle!(sc,network,freq)

Plots the the source stability circle on Smith Chart sc from network at frequency freq.

source

Marconi.plotSmithCircle! — Method.

plotSmithCircle!(sc, xc, yc, rad)

Plots a cricle with center coordinates (xc,yc) on the $\Gamma$ plane with radius rad on an existing Smith Chart object.

Returns the sc object

source

Marconi.plotSmithData! — Method.

plotSmithData!(sc, network,(1,1))

Plots the S(1,1) parameter from network on an existing Smith Chart sc

Returns the sc object

source

Marconi.plotSmithData — Method.

plotSmithData(network,(1,1))

Plots the S(1,1) parameter from network on a Smith Chart.

Returns a PGFPlotsX.SmithChart object.

source

Marconi.plotVSWR! — Method.

plotVSWR!(sc,VSWR)

Plots the circle that represents a VSWR of VSWR onto an existing Smith Chart.

source

Marconi.readTouchstone — Method.

readTouchstone("myFile.sNp")

Reads the contents of myFile.sNp into a Network object. This will convert all file types to S-Parameters, Real/Imaginary

Currently does not support reference lines (Different port impedances) or noise parameters

source

Marconi.s2t — Method.

    s2t(s)

Converts S-Parameters s to T-Parameters.

source

Marconi.s2z — Method.

    s2z(s)

Converts S-Parameters s to Z-Parameters. Optionally include reference impedance with kwarg Z0 with s2z(s,Z0=50).

source

Marconi.t2s — Method.

    t2s(t)

Converts T-Parameters t to S-Parameters.

source

Marconi.testDelta — Method.

testDelta(network)

Returns a vector of Δ, the determinant of the scattering matrix. Optionally, returns Δ for S-Parameters at position pos.

source

Marconi.testK — Method.

testK(network)

Returns a vector of the magnitude of K, the Rollet stability factor.

source

Marconi.testMagDelta — Method.

testMagDelta(network)

Returns a vector of Δ, the determinant of the scattering matrix. Optionally, returns |Δ| for S-Parameters at position pos.

source

Marconi.y2s — Method.

    y2s(y)

Converts Y-Parameters y to S-Parameters. Optionally include reference impedance with kwarg Z0 with y2s(y,Z0=50).

source

Marconi.z2s — Method.

    z2s(z)

Converts Z-Parameters z to S-Parameters. Optionally include reference impedance with kwarg Z0 with z2s(z,Z0=50).

source

Marconi.Γ — Method.

Γ(Z,Z0)

Calculates the complex reflection coefficient Γ from impedance Z. Z0 is optional and defaults to 50.

source

Marconi.∠ — Method.

∠(mag,angle)

A nice compact way of representing phasors. Angle is in degrees.

source

Marconi.processTouchstoneLine — Method.

Internal function to process touchstone lines

source