File IO

File IO

Marconi.jl supports reading and writing standard Touchstone files with most of Touchstone spec implemented including n-ports, non standard port impedances, and S, Y, Z, H, and G parameter reading.

Marconi.jl currently does not support noise parameters nor per-port impedance.

Functions

Marconi.readTouchstone — Function.

readTouchstone("myFile.sNp")

Reads the contents of myFile.sNp into a Network object. This will convert all file types to S-Parameters, Real/Imaginary

Currently does not support reference lines (Different port impedances) or noise parameters

source