Network Analysis

Network Analysis

Marconi.jl supports conversion between S, Y, Z, H, G, ABCD and T parameters.

There are functions to cascade networks and deembed networks as well.

Functions

Marconi.y2s — Function.

    y2s(y)

Converts Y-Parameters y to S-Parameters. Optionally include reference impedance with kwarg Z0 with y2s(y,Z0=50).

Example

y11 =  0.0488133074245012 - 0.390764155450191im
y12 = -0.0488588365420561 + 0.390719345880018im
y21 = -0.0487261119282660 + 0.390851884427087im
y22 =  0.0487710062903760 - 0.390800401433241im

y_params = [y11 y12; y21 y22]

julia> y2s(y_params)
2×2 Array{Complex{Float64},2}:
 0.00381839+0.0247966im    0.996111-0.0249991im
   0.996392-0.0253812im  0.00374364+0.0249161im

source

Marconi.z2s — Function.

    z2s(z)

Converts Z-Parameters z to S-Parameters. Optionally include reference impedance with kwarg Z0 with z2s(z,Z0=50).

Example

z11 = -14567.2412789287 - 148373.315116592im
z12 = -14588.1106171651 - 148388.583516562im
z21 = -14528.0522132692 - 148350.705757767im
z22 = -14548.5996561832 - 148363.457002006im

z_params = [z11 z12; z21 z22]

julia> z2s(z_params)
2×2 Array{Complex{Float64},2}:
 0.00381839+0.0247966im    0.996392-0.0253812im
   0.996111-0.0249991im  0.00374364+0.0249161im

source

Marconi.s2z — Function.

    s2z(s)

Converts S-Parameters s to Z-Parameters. Optionally include reference impedance with kwarg Z0 with s2z(s,Z0=50).

source

Examples