Plotting

Plotting

Note

The plotting library depends on a working installation of PGFPlotsX.jl

Currently, one can plot networks on

Smith Charts

Rectangular Plots

Polar Plots

Circles on Smith Charts

Gain Circles

Q Circles

VSWR Circles

Constant-R Circles

Constant-G Circles

Constant-X Circles

Constant-Y Circles

Functions

Marconi.plotSmith — Function.

plotSmith(network,(1,1))

Plots the S(1,1) parameter from network on a Smith Chart.

Returns a PGFPlotsX.SmithChart object.

source

Marconi.plotSmith! — Function.

plotSmith!(sc, network,(1,1))

Plots the S(1,1) parameter from network on an existing Smith Chart sc

Returns the sc object

source

Marconi.plotSmithCircle! — Function.

plotSmithCircle!(sc, xc, yc, rad)

Plots a cricle with center coordinates (xc,yc) on the $\Gamma$ plane with radius rad on an existing Smith Chart object.

Returns the sc object

source