Public

Public Functions for Marconi

Index

Marconi.Network

Marconi.plotSmith

Marconi.plotSmith!

Marconi.plotSmithCircle!

Marconi.processTouchstoneLine

Marconi.readTouchstone

Marconi.s2z

Marconi.y2s

Marconi.z2s

Public Interface

Marconi.plotSmith! — Method.

plotSmith!(sc, network,(1,1))

Plots the S(1,1) parameter from network on an existing Smith Chart sc

Returns the sc object

source

Marconi.plotSmith — Method.

plotSmith(network,(1,1))

Plots the S(1,1) parameter from network on a Smith Chart.

Returns a PGFPlotsX.SmithChart object.

source

Marconi.plotSmithCircle! — Method.

plotSmithCircle!(sc, xc, yc, rad)

Plots a cricle with center coordinates (xc,yc) on the $\Gamma$ plane with radius rad on an existing Smith Chart object.

Returns the sc object

source

Marconi.readTouchstone — Method.

readTouchstone("myFile.sNp")

Reads the contents of myFile.sNp into a Network object. This will convert all file types to S-Parameters, Real/Imaginary

Currently does not support reference lines (Different port impedances) or noise parameters

source

Marconi.s2z — Method.

    s2z(s)

Converts S-Parameters s to Z-Parameters. Optionally include reference impedance with kwarg Z0 with s2z(s,Z0=50).

source

Marconi.y2s — Method.

    y2s(y)

Converts Y-Parameters y to S-Parameters. Optionally include reference impedance with kwarg Z0 with y2s(y,Z0=50).

Example

y11 =  0.0488133074245012 - 0.390764155450191im
y12 = -0.0488588365420561 + 0.390719345880018im
y21 = -0.0487261119282660 + 0.390851884427087im
y22 =  0.0487710062903760 - 0.390800401433241im

y_params = [y11 y12; y21 y22]

julia> y2s(y_params)
2×2 Array{Complex{Float64},2}:
 0.00381839+0.0247966im    0.996111-0.0249991im
   0.996392-0.0253812im  0.00374364+0.0249161im

source

Marconi.z2s — Method.

    z2s(z)

Converts Z-Parameters z to S-Parameters. Optionally include reference impedance with kwarg Z0 with z2s(z,Z0=50).

Example

z11 = -14567.2412789287 - 148373.315116592im
z12 = -14588.1106171651 - 148388.583516562im
z21 = -14528.0522132692 - 148350.705757767im
z22 = -14548.5996561832 - 148363.457002006im

z_params = [z11 z12; z21 z22]

julia> z2s(z_params)
2×2 Array{Complex{Float64},2}:
 0.00381839+0.0247966im    0.996392-0.0253812im
   0.996111-0.0249991im  0.00374364+0.0249161im

source

Marconi.Network — Type.

The base Network type for representing n-port linear networks with characteristic impedance Z0. By default, the network is stored as S-Parameters with the corresponding frequency list.

source

Marconi.processTouchstoneLine — Method.

Internal function to process touchstone lines

source