Public

Public Functions for Marconi

Index

Marconi.DataNetwork

Marconi.plotSmithCircle!

Marconi.plotSmithData

Marconi.plotSmithData!

Marconi.processTouchstoneLine

Marconi.readTouchstone

Marconi.s2z

Marconi.y2s

Marconi.z2s

Public Interface

Marconi.DataNetwork — Type.

The base Network type for representing n-port linear networks with characteristic impedance Z0. By default, the network is stored as S-Parameters with the corresponding frequency list.

source

Marconi.plotSmithCircle! — Method.

plotSmithCircle!(sc, xc, yc, rad)

Plots a cricle with center coordinates (xc,yc) on the $\Gamma$ plane with radius rad on an existing Smith Chart object.

Returns the sc object

source

Marconi.plotSmithData! — Method.

plotSmithData!(sc, network,(1,1))

Plots the S(1,1) parameter from network on an existing Smith Chart sc

Returns the sc object

source

Marconi.plotSmithData — Method.

plotSmithData(network,(1,1))

Plots the S(1,1) parameter from network on a Smith Chart.

Returns a PGFPlotsX.SmithChart object.

source

Marconi.readTouchstone — Method.

readTouchstone("myFile.sNp")

Reads the contents of myFile.sNp into a Network object. This will convert all file types to S-Parameters, Real/Imaginary

Currently does not support reference lines (Different port impedances) or noise parameters

source

Marconi.s2z — Method.

    s2z(s)

Converts S-Parameters s to Z-Parameters. Optionally include reference impedance with kwarg Z0 with s2z(s,Z0=50).

source

Marconi.y2s — Method.

    y2s(y)

Converts Y-Parameters y to S-Parameters. Optionally include reference impedance with kwarg Z0 with y2s(y,Z0=50).

source

Marconi.z2s — Method.

    z2s(z)

Converts Z-Parameters z to S-Parameters. Optionally include reference impedance with kwarg Z0 with z2s(z,Z0=50).

source

Marconi.processTouchstoneLine — Method.

Internal function to process touchstone lines

source